requirements.txt
# tokenization
pygments==2.10.0
tokenizers==0.9.4
transformers==4.2.1

# utils
pandas==1.3.1
hydra-core==1.1.1

# code quality
black
mypy
pytest
